Day 4 done
Got part 1 done naively in rust.  Got a bit stuck on part 2, I really need to reach more instinctively for recursion in rust.  I do it so naturally in python, but the friction in rust seems to make my brain not even consider it.  I need to smooth that out.